+++ src/instrumentation.ts
@@ -2,9 +2,7 @@ import { registerOTel } from '@vercel/otel';
 
 export async function register() {
     if (process.env.NEXT_RUNTIME === 'nodejs') {
-        await import('./db/create');
-        await import('./db/migrate');
-        await import('./db/seed');
+        await import('./db/seeder');
     }
 
     registerOTel({ serviceName: 'halcyon-nextjs' });
